move onClickEnable on screen
