+++ src/screens/onboarding-biometry.tsx
@@ -1,11 +1,13 @@
-import React, {useCallback} from 'react';
+import React, {useCallback, useState} from 'react';
 
 import {OnboardingBiometry} from '@app/components/onboarding-biometry';
-import {useTypedNavigation, useTypedRoute} from '@app/hooks';
+import {useApp, useTypedNavigation, useTypedRoute} from '@app/hooks';
 
 export const OnboardingBiometryScreen = () => {
   const navigation = useTypedNavigation();
   const route = useTypedRoute<'onboardingBiometry'>();
+  const app = useApp();
+  const [error, setError] = useState('');
   const {biometryType} = route.params;
 
   const onClickSkip = useCallback(() => {
@@ -15,7 +17,23 @@ export const OnboardingBiometryScreen = () => {
     });
   }, [route, navigation]);
 
+  const onClickEnable = useCallback(async () => {
+    try {
+      await app.biometryAuth();
+      app.biometry = true;
+      onClickSkip();
+    } catch (e) {
+      if (e instanceof Error) {
+        setError(e.message);
+      }
+    }
+  }, [app, onClickSkip]);
+
   return (
-    <OnboardingBiometry onClickSkip={onClickSkip} biometryType={biometryType} />
+    <OnboardingBiometry
+      onClickSkip={onClickSkip}
+      onClickEnable={onClickEnable}
+      error={error}
+      biometryType={biometryType} />
   );
 };
